Added ablity to specify new ID or Name in SDK migration call
Review @95770
"[PATCH 4/4] Added ablity to specify new ID or Name in SDK
migration call and prlctl."
Reviewed-by: myakhin
Reviewed-by: shrike
Observers: sergeyt, dim, igor

This seria of patches introduces new ability to change Uuid or
Name of a VM and ID or Name of a CT during migration. For this
purpose protocol was extended with new parameter that can
contain either ID or Name of VE. This new parameter is read
out from migration command itself and is transferred between
migration Source and Target servers as part of check-preconditions
command.

This patch implements ablity to specify new ID or Name in SDK
migration call and prlctl. In case of prlctl the usage is pretty
straight-forward, e.g.:
prlctl migrate 127.0.0.1/300 10.30.110.100/600
Issues:
Affects Critical Issue PSBM-26671 - "Implement remote clone operation in SDK."