Handle backup attached to a Windows VM in PrlVmCfg_AddDefaultDeviceEx
see #PSBM-30201 @100642
Review @100642

Reviewed-by: myakhin
Observers: igor
A special case of a backup, attached to a Windows VM, is handled by
PrlVmCfg_AddDefaultDeviceEx() API call. There we need to take into
account the positions of already existing bootable HDDs and place
the newly attached backup after the "last" bootable HDD, according
to the Windows boot preference order (IDE->SATA->SCSI).
This is due to the following issue (#PSBM-30201): if Windows sees
several clones of the same disk during the boot ( backup of the
Windows system disk was attached to the same VM), then it first
tries to boot from the disk, connected to IDE interface, then tries
SATA, and in the end SCSI. So we have to place an attached backup
after the last bootable HDD, in order to not to break Windows booting.
Tests for the new functionality will be added in the 2nd patch.
Issues:
Affects Critical Issue PSBM-30201 - "[Buse] Windows guest can BSOD
if sata position of backup disk if less then base disk."