PySDK: Forcibly convert job statuses to signed int
PSBM-31263

It was found out that specification of big signed decimal values inside
C++ enums leads to platform specific behaviour (and even compiler
specific). These values can be either signed or unsigned after
compilation.
(more detailed description can be found here:
http://www.open-std.org/jtc1/sc22/wg21/docs/papers/2004/n1719.pdf
paragraph 2.2.2)

Unfortunately it leads to problems in our Python SDK. When we add
constant to it we take values as is. When we retrieve values from
PrlJob_GetStatus funstion it converted to signed integer.

So the following absolutely correct code will not work because the
values are different:
job = srv.login("127.0.0.1", "root", "1q2w3e")

if job.get_status() == consts.PJS_RUNNING:
    print "Running"

Good solution for this problem was introduced in C++11. It
allows to specify in enum declaration which type to use. However it
is unsafe to make change in SDK constants (due to
compatibility problems) thus lets introduce a simple
workaround and forcibly convert all SDK Job statuses to
signed integer before their addition to Python module.