Add PRL_VIRTUAL_NETWORK_FLAGS enum for PrlSrv_AddVirtualNetwork
There only one flag is going to be used now: PRL_USE_VNET_NAME_FOR_BRIDGE_NAME.
It means that when a virtual network is created with this flag specified,
a function that takes actual action shouldn't create a new bridge for it,
rather it should use existing bridge with the name of Virtual Network.
On deletion such networks, bridges assigned to them shouldn't be destroyed
automatically.

Signed-off-by: Maxim Nestratov <mnestratov@parallels.com>