MERGE Implement detach and reattach for the client side of the IOCommunication connection (#PSBM-29120)
Previously, only the server side of the IOCommunication connection could
be detached and then reattached. Implement the same functionality for
the client side, so that both sides of the connection could be detached
simultaneously. Also, add an ability to specify which side of the
connection (client or server) is being reattached.
This functionality is needed for a BUSE plugin, which will use the
established Dispatcher connection to read backup data from a remote
server.

Signed-off-by: Igor Bazhitov <ibazhitov@virtuozzo.com>