Add a specific error for cases when a new private path is given when restoring backup over an existing VM. #PSBM-45427
When we try to restore a backup for an existing VM, giving a non-empty sTargetVmHomePath and an empty sTargetVmName to PrlSrv_RestoreVmBackup, the dispatcher fails with a generic error PRL_ERR_BAD_PARAMETERS, which may confuse customers when they are trying to run e.g. "prlctl restore -t <uuid> --dst <path>". This patch introduces a separate error code for this error.

Signed-off-by: Anton Sanarov <asanarov@virtuozzo.com>