Output a better error message when the connection to the local dispatcher is lost. #PSBM-47281
Now, the message is like this:
Do you want to reconnect to the server "/var/run/prl_disp_service.socket"? The connection to the server has been lost. You need to reconnect to it to continue working with your virtual machines.

It is looking rather ugly, and may confuse users. This patch changes the server name to "localhost" instead of the socket path.

Signed-off-by: Anton Sanarov <asanarov@virtuozzo.com>