Split SDK headers to two parts - basic and the rest
We need to extract basic definitions (errors, types, etc.) to a separate
package, because they are used in every prl* project, while the rest is not
used and should not be used.

This split will break a circular dependency issue, which was introduced during
the not-well-thought divorce of the initial single repository.