network: add new new network adapter type PNA_BRIDGE
Also:
 - rename PNA_BRIDGED_ETHERNET to PNA_BRIDGED_NETWORK,
 - remove commented out value

Intention of this change is the following. Currently a virtual network
interface of PNA_BRIDGED_ETHERNET type is nothing but an interface connected
with bridge via some Network. Let's call it accordingly. Anyway, old name
PNA_BRIDGED_ETHERNET is left for compatibility.
On the other hand, connecting a virtual network adapter directly to some bridge
should be made via some special type of interface. Let it be PNA_BRIDGE.