bug #PSBM-59525 aka bused killed by SIGSEGV
Q_GLOBAL_STATIC creates an instance for each call on a first access
to the singleton storage. then one instance gets into the storage
the rest are destroyed. we cleaned the storage from the destructor
in assumption that the only object of the kind is being destroyed
which is not the case when an extra instance is being destroyed.

let's rely on Q_GLOBAL_STATIC internals to destroy the CStdinMaintainer
global instance and track only thread termination from the
g_pThreadsDestructor. also terminate the CStdinMaintainer thread
from the application's aboutToQuit slot instead of a dedicated
external function.