bug #PSBM-59918 aka prlctl exec killed by SIGSEGV
qwaitcondition wait quits immediately when the mutex is recursive.
thus we don't wait for a finish of all the pending threads.
make the mutex not recursive and handle finished signals using a
direct connection to operate in a callback like way because the
processor is ready to operate in a concurrent environment. also
we need to wait explicitly for each tracked thread because qthread
performs more actions after the finished signal emmission. thus
it is unsafe to delete a qthread object immediately on receipt of
the signal.