bug #PSBM-61204 aka vzt-psbm-validation: vm_resume ERROR: dump: core-snmpd-1488228561.9053
PRL_OBJECT_BY_HANDLE inside the loop can steal the last handle
instance reference. destruction of the reference in the end of
the loop body incurs a call UnregisterInstance from the handle
destructor. the call doesn't hang on the g_ControlValidityMapMutex
because the later is recursive. it updates g_ControlValidityMap
resulting in the _data_it iterator invalidation. ub
let's move the loop out of the lock scope and iterate over a
copy of the map content. this should preserve consistency of
the map.