bug #PSBM-67887 aka vzt-psbm-validation: lin.vm_install TIMEDOUT: Function wait timeouted
a socketclient a) emits signal for every response package; b) stores
that package in an associated io job. the pve control doesn't poll
io job and the progress sticks because the io job capacity gets
exhausted.
let's take one package from the io joib on receipt of a response
package to free space for another one.