bug #PSBM-79714 aka "net-snmp" crash: "snmpd killed by SIGSEGV"
SDK handle id values can overlap. we insert a value for a handle
into the handles map without testing if the value is in the map
already. this leads to a situation when we replace a handle in the
map with the one potentially of another type.
let's make a test for uniqueness of a handle id and also check if
a device vm handle is the same as the one which constructs the
handle.