merge from VZ6: expose SPEC_CTRL to guests if supported @nowidth
Intel processors that include the mitigation for Spectre #2
vulnerability announce the presence of the necessary features in
register EDX of the CPUID leaf 0x7.

Make sure the guests can see and use these features, while handling live
migration correctly.

This patch doesn't make the monitor itself use the SPEC_CTRL msr so it
remains vulnerable; this will come in a followup patch.