bug #PVA-37383 aka pvaagent on node is using 150Gb RAM
first sdk creates a vm config with a random uuid and registers that
uuid in a map for unregistering inside destructor. then new config
is assigned to the vm handle with another uuid, that uuid is also
registered in the map but random uuid stays in the map because
nobody calls unregister and worse nobody will. finally the map is
full of untracked records which are not leaks but still consume
resources.
unregister old uuid before registering new one.