bug #PSBM-92425 aka prlsrvctl net list crashed
seems we have handles destroyed when the handles is empty but the
handles list for cleanup is not freed. we destroy a handle object
inside a release call and end with a double free.
let's mark handles which are outside of the map and make them
eligible for an explict destruction inside release member.